优化代码以及添加其他命令业务层代实现

+++ zhuozun-websocket-server/src/main/java/com/qs/game/common/KunGold.java
@@ -6,7 +6,7 @@ package com.qs.game.common;
 public interface KunGold {
 
 
-    static int goldByType(int kunType) {
+    static long goldByType(int kunType) {
         switch (kunType) {
             case KunType.TYPE_1:
                 return 10;
